+++ src/main/java/explorableviz/authoringassistant/FluidCLI.java
@@ -1,13 +1,8 @@
-package explorableviz.transparenttext;
+package explorableviz.authoringassistant;
 
 import org.apache.commons.io.FileUtils;
 import java.io.File;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.logging.Logger;
