+++ src/main/java/explorableviz/authoringassistant/InContextLearning.java
@@ -1,6 +1,6 @@
-package explorableviz.transparenttext;
+package explorableviz.authoringassistant;
 
-import explorableviz.transparenttext.paragraph.Expression;
+import explorableviz.authoringassistant.paragraph.Expression;
 import it.unisa.cluelab.lllm.llm.prompt.PromptList;
 import kotlin.Pair;
 
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
-import static explorableviz.transparenttext.Program.loadPrograms;
+import static explorableviz.authoringassistant.Program.loadPrograms;
 
 public class InContextLearning {
     private final String systemPrompt;
