+++ src/main/java/explorableviz/authoringassistant/Main.java
@@ -1,22 +1,9 @@
-package explorableviz.transparenttext;
+package explorableviz.authoringassistant;
 
 import kotlin.Pair;
 
-import explorableviz.transparenttext.Program.QueryResult;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.w3c.dom.Node;
+import explorableviz.authoringassistant.Program.QueryResult;
 
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
-import javax.xml.xpath.XPath;
-import javax.xml.xpath.XPathConstants;
-import javax.xml.xpath.XPathFactory;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
