+++ src/main/java/explorableviz/authoringassistant/llm/FluidOpenAIGpt4oAgent.java
@@ -1,4 +1,4 @@
-package explorableviz.transparenttext.llm;
+package explorableviz.authoringassistant.llm;
 
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.theokanning.openai.completion.chat.ChatFunction;
@@ -6,7 +6,7 @@ import com.theokanning.openai.completion.chat.ChatFunctionCall;
 import com.theokanning.openai.completion.chat.ChatMessage;
 import com.theokanning.openai.service.FunctionExecutor;
 import com.theokanning.openai.service.OpenAiService;
-import explorableviz.transparenttext.paragraph.Expression;
+import explorableviz.authoringassistant.paragraph.Expression;
 import it.unisa.cluelab.lllm.llm.agents.generic.OpenAIEvaluatorAgent;
 import it.unisa.cluelab.lllm.llm.prompt.Prompt;
 import org.jetbrains.annotations.NotNull;
