+++ src/main/java/explorableviz/authoringassistant/variable/Variables.java
@@ -1,4 +1,4 @@
-package explorableviz.transparenttext.variable;
+package explorableviz.authoringassistant.variable;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
